# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TamperRequest": {
        "additionalProperties": false,
        "description": "Which block to corrupt, and how.",
        "properties": {
          "block_index": {
            "minimum": 0.0,
            "title": "Block Index",
            "type": "integer"
          },
          "mode": {
            "default": "payload",
            "pattern": "^(payload|hash|timestamp)$",
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "block_index"
        ],
        "title": "TamperRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Autonomous revenue recovery agent for Razorpay merchants. Bounded agent: the LLM diagnoses and writes; it never touches money.",
    "title": "RevPilot AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/audit/ledger": {
      "get": {
        "description": "Page through the chain. Read-only; no endpoint updates or deletes.",
        "operationId": "list_ledger_api_v1_audit_ledger_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Ledger Api V1 Audit Ledger Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read the ledger",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/v1/audit/tamper": {
      "post": {
        "description": "Deliberately corrupt one block.\n\nGated on ``settings.simulation_allowed`` \u2014 an endpoint that damages the\naudit log must be unreachable in production, and the check is on the\nenvironment rather than on a header a caller controls.",
        "operationId": "tamper_api_v1_audit_tamper_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TamperRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Tamper Api V1 Audit Tamper Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "[dev only] Corrupt a block so the verifier can be seen catching it",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/v1/audit/verify": {
      "get": {
        "description": "Recompute every block from its own stored fields.\n\nNothing stored is trusted \u2014 not ``current_hash``, not ``payload_hash``, not\nthe link. The response names the specific failure and the index it occurred\nat, rather than a bare boolean, because \"invalid\" is not actionable.",
        "operationId": "verify_chain_api_v1_audit_verify_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Verify Chain Api V1 Audit Verify Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recompute the audit chain and report any divergence",
        "tags": [
          "audit"
        ]
      }
    },
...
  }
}